cluster/config: move subnet parsing to utils
